Move sampling to a separate file.
