+++ metaphor/bigquery/profile/config.py
@@ -4,7 +4,7 @@ from typing import Optional
 from serde import deserialize
 
 from metaphor.bigquery.config import BigQueryRunConfig
-from metaphor.common.extractor import SamplingConfig
+from metaphor.common.sampling import SamplingConfig
 
 
 @deserialize
